# KDT_HACK/src/main/java/com/hackathon/KDT_HACK/Registration/ChangePasswordRequest.java
package com.hackathon.KDT_HACK.Registration;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class ChangePasswordRequest {
    @NotBlank(message = "Старый пароль обязателен")
    private String oldPassword;

    @NotBlank(message = "Новый пароль обязателен")
    @Size(min = 8, max = 100, message = "Пароль должен содержать от 8 до 100 символов")
    private String newPassword;
}