# swagger.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/secured/user/{id}":{"get":{"tags":["main-controller"],"operationId":"userAccess","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"post":{"tags":["main-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/secured/user/{id}/change-password":{"post":{"tags":["main-controller"],"operationId":"changePassword","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/signup":{"post":{"tags":["security-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/signin":{"post":{"tags":["security-controller"],"operationId":"signin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/admin":{"get":{"tags":["admin-controller"],"operationId":"admin","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["admin-controller"],"operationId":"httpStatus","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Skills"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 EARLY_HINTS","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","303 SEE_OTHER","304 NOT_MODIFIED","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 CONTENT_TOO_LARGE","413 PAYLOAD_TOO_LARGE","414 URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","421 MISDIRECTED_REQUEST","422 UNPROCESSABLE_CONTENT","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]}}}}}}},"/judge":{"get":{"tags":["admin-controller"],"operationId":"judge","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/schedule/{id}":{"get":{"tags":["schedule-controller"],"operationId":"getEventById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Schedule"}}}}}}},"/api/schedule":{"get":{"tags":["schedule-controller"],"operationId":"getAllSchedules","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Schedule"}}}}}}}},"/api/schedule/admin":{"get":{"tags":["schedule-controller"],"operationId":"getAllSchedules_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Schedule"}}}}}}}}},"components":{"schemas":{"SkillDto":{"type":"object","properties":{"skillId":{"type":"integer","format":"int64"},"level":{"type":"integer","format":"int32"}},"required":["level","skillId"]},"UpdateUserRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":2,"pattern":"^[\\p{L}\\s\\-']+$"},"lastName":{"type":"string","maxLength":100,"minLength":2,"pattern":"^[\\p{L}\\s\\-']+$"},"username":{"type":"string"},"birthday":{"type":"string","format":"date"},"email":{"type":"string"},"phone":{"type":"string","pattern":"^\\+79\\d{9}$"},"telegram":{"type":"string","pattern":"^https://t\\.me/[a-zA-Z0-9][a-zA-Z0-9_]{4,31}$"},"job":{"type":"string","enum":["FRONT","BACK","DESIGNER","PROJECT","GAME"]},"level":{"type":"string","enum":["BEGINNER","INTERMEDIATE","ADVANCED"]},"skills":{"type":"array","items":{"$ref":"#/components/schemas/SkillDto"}}},"required":["birthday","email","lastName","name","username"]},"Skills":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"category":{"type":"string","enum":["SOFT","HARD"]}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"telegram":{"type":"string"},"birthday":{"type":"string","format":"date"},"password":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN","JUDGE","PARTNER"]},"status":{"type":"string","enum":["ACTIVE","BLOCKED","DELETED"]},"job":{"type":"string","enum":["FRONT","BACK","DESIGNER","PROJECT","GAME"]},"level":{"type":"string","enum":["BEGINNER","INTERMEDIATE","ADVANCED"]},"skills":{"type":"array","items":{"$ref":"#/components/schemas/UserSkill"}}}},"UserSkill":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"skill":{"$ref":"#/components/schemas/Skills"},"level":{"type":"integer","format":"int32","maximum":10,"minimum":1}}},"ChangePasswordRequest":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string","maxLength":100,"minLength":8}},"required":["newPassword","oldPassword"]},"SignupRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":2,"pattern":"^[\\p{L}\\s\\-']+$"},"lastname":{"type":"string","maxLength":100,"minLength":2,"pattern":"^[\\p{L}\\s\\-']+$"},"username":{"type":"string"},"birthday":{"type":"string","format":"date"},"email":{"type":"string"},"password":{"type":"string","maxLength":100,"minLength":8}},"required":["birthday","email","lastname","name","password","username"]},"SigninRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"JwtResponse":{"type":"object","properties":{"token":{"type":"string"}}},"Schedule":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","maxLength":2147483647,"minLength":5},"briefDescription":{"type":"string","maxLength":2147483647,"minLength":5},"description":{"type":"string","maxLength":2147483647,"minLength":5},"date":{"type":"string","format":"date-time"}},"required":["briefDescription","date","description","name"]}}}}